requirements.txt
hyperliquid-python-sdk>=0.7.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
eth-account>=0.10.0
requests>=2.31.0
websockets>=12.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.18.0


